Upgraded home page.
I created 3 more objects and uploaded their respective images.
Modified the home() in Project/views.py to retrieve all the Project objects from the
database and pass it to Project/Templetes/Project/home.html.
Finally added necessary logic to display all the details for every Project object in the
list passed from home().
This was just for demo and will be changed.